cli: Turn advisory list into template
Thiw way the same output function can by used by both dnf5 and
dnf5daemon-client.
